Change track updating logic

+++ src/main/java/me/fizzika/tankirating/dto/alternativa/track/AlternativaUsageTrack.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public abstract class AlternativaUsageTrack extends AlternativaTrackEntity {
 
-    private long usages;
+    private int usages;
 
 }
